test(ui): keep metadata regression within suite limits
